requirements/development.txt
-r common.txt

coverage==7.6.*
django_extensions==3.2.*
django-debug-toolbar==4.4.*
ipdb==0.13.13
psycopg[binary,pool]==3.2.*
pylint-django==2.5.*
pytest==8.3.*
pytest-cov==5.0.*
pytest-django==4.8.*
pytest-html==4.1.*
pytest-order==1.2.*
pytest-env==1.1.*
pytest-sugar==1.0.*
ruff==0.5.*
